Séparation de locales/Liste.tsx en deux composants
